Refactor cassandra-quota-mailing: using the QuotaDTO event and related classes from quota-mailing module
